added endpoints to server.js
