+++ .gitignore
@@ -2,4 +2,3 @@
 node_modules
 config.js
 test.js
-./dist
