+++ controllers/kataCtrl.js
@@ -1,0 +1,27 @@
+const app = require('../server');
+const db = app.get('db');
+
+module.exports = {
+    getKata: (res, req, next) => {
+
+    },
+
+	postSolution: (res, req, next) => {
+
+    },
+
+    getRandomKata: (res, req, next) => {
+
+    },
+
+    getCompletedKatas: (res, req, next) => {
+
+    },
+
+    getKataSolutions: (res, req, next) => {
+
+    },
+
+    
+
+}
